feat: test for register color in CLI
